Add groups.
Now parse '[group]' elements. Assignments that precede all groups
are put into same top-level dict as the groups. Assignments that
follow a group are put into a dict which is put into the top-level
dict under the group name.

System test now failing on a string value with embedded "\n".